Tweak warning and spelling issue #7993
Fix spelling of headers in the output of `condor_userprio` and
a compiler warning noticed on Mac OS X.

+++ src/condor_negotiator.V6/matchmaker.cpp
@@ -1578,7 +1578,7 @@ negotiationTime ()
 							NegotiatorInterval);
 		dprintf(D_FULLDEBUG,
 			"New cycle requested but just finished one -- delaying %lu secs\n",
-			cycle_delay - elapsed);
+			static_cast<long unsigned>(cycle_delay - elapsed));
 		return;
 	}
